saving changes to app ternary
