+++ client/src/components/MiniCardContainer/MiniCardContainer.js
@@ -1,20 +1,17 @@
 import React, {useContext, useState, useEffect} from "react";
 import Minikidcard from "../Minikidcard/Minikidcard";
-import UserContext from "../../utils/userContext";
 
-function MiniCardContainer() {
-    const {dbUser}=useContext(UserContext);
+function MiniCardContainer(props) {
 
   return (
     <div>
       {/* This should be mapped */}
-      {dbUser.child ? (
-        dbUser.child.map((data) => (
+      {props.kidList ? (
+        props.kidList.map((data) => (
           <Minikidcard key={data._id} name={data.kidname} age={data.kidage} gender={data.gender} />
         ))
       ) : (
-        <p>
-            ""</p>
+      null
       )}
     </div>
   );
