+++ controllers/userController.js
@@ -48,7 +48,7 @@ module.exports = {
   },
   addKidByEmail: function(req, res) {
     db.User
-      .findOneAndUpdate({ email: req.params.email }, {$push: req.body.push})
+      .findOneAndUpdate({ email: req.params.email }, {$push: req.body.push}, {new: true})
       .then(dbModel => res.json(dbModel))
       .catch(err => res.status(422).json(err));
   }
